fix(lyrics)!: URL encoding ('%20' -> '+')
LRCLIB implemented a change in their API, causing the space characters
in song, artist, and album names to change into '+' characters instead
of the standard '%20' encoding. This commit fixes that.

Also, instead of checking the http header for specific character
combinations (e.g. 400, 404, etc.), we now get the http response code
via curl directly, and check that instead.

Also, removed all the now-redundant pid-checking at the bottom of the
function since the file writing is now atomic.

BREAKING CHANGE: The script won't work with APIs that switch out spaces
to '%20's instead of '+' characters.

+++ modules/scripts/payload/waybar-lyrics.sh
@@ -22,29 +22,27 @@ export DEV_ENV='/home/jaysh/dev'
 source "${DEV_ENV}/lib/urlencode" && \
     export -f urlencode
 
-cleanup_http_header() {
-    local header="$1"
-    local unwanteds='server|date|content-type|content-length|connection|vary'
-    local err_codes='400|404|402|429|500|502|503|504'
-    grep -viE "$unwanteds" "$header" | \
-        grep -iE "$err_codes" - | \
-        sponge "$header"
-    return
-}
-
 api_call() {
     local uri="$1"
-    local api_resp_file="$2"
-    local lyrics_file="$3"
+    local lyrics_file="$2"
+
     coproc 'API_PIPELINE' {
-        curl -s "$uri" --dump-header "$api_resp_file" | \
-            jq -r '.plainLyrics // ""' | \
-            iconv -t ASCII//TRANSLIT | \
-            sed 's/^[ \t]*//' | \
-            sed '/./,$!d' | \
-            cat -s - | \
-            fold -s -c -w 34 - | \
-            pr -t -T -c2 -w 71 -l 100 -S' | ' - > "$lyrics_file"
+        local raw_resp="${lyrics_file}.raw"
+        local tmp_file="${lyrics_file}.tmp"
+        local http_code=$(curl -s -w "%{http_code}" -o "$raw_resp" "$uri")
+        if [[ "$http_code" == "200" ]]; then
+            jq -r '.plainLyrics // ""' "$raw_resp" | \
+                iconv -t ASCII//TRANSLIT | \
+                sed 's/^[ \t]*//' | \
+                sed '/./,$!d' | \
+                cat -s - | \
+                fold -s -c -w 34 | \
+                pr -t -T -c2 -w 71 -l 100 -S' | ' > "$tmp_file"
+            mv -f "$tmp_file" "$lyrics_file"
+        else
+            echo -e "No lyrics found.\n${uri}" > "$lyrics_file"
+        fi
+        rm -f "$raw_resp"
     }
     REPLY="$API_PIPELINE_PID"
 }
@@ -58,12 +56,14 @@ api_call() {
 #   duration (s) = 311
 get_lyrics() {
     local artist_name=$(urlencode "$1")
+    artist_name=${artist_name//%20/+}
     local track_name=$(urlencode "$2")
+    track_name=${track_name//%20/+}
     local album_name=$(urlencode "$3")
+    album_name=${album_name//%20/+}
     local duration=$4 # this is already given in seconds
 
     local api_pid_file="${LYRICS_D}/.${artist_name}-${track_name}.pid.id"
-    local http_header="${LYRICS_D}/.${artist_name}-${track_name}.http.header"
     local lyrics_file="${LYRICS_D}/${artist_name}-${album_name}-${track_name}.lyrics"
     local fetching_msg='Fetching lyrics...'
     local uri="https://lrclib.net/api/get?artist_name=${artist_name}&track_name=${track_name}&album_name=${album_name}&duration=${duration}"
@@ -72,25 +72,16 @@ get_lyrics() {
     if [[ -f "$api_pid_file" ]]; then
         api_pid=$(cat "$api_pid_file")
         if ps -p "$api_pid" > /dev/null 2>&1; then
-            echo "$fetching_msg" > "$lyrics_file"
+            if [[ ! -s "$lyrics_file" ]]; then
+                echo "$fetching_msg" > "$lyrics_file"
+            fi
             return
         else
             rm -f "$api_pid_file"
             return
         fi
     elif [[ ! -f "$lyrics_file" ]]; then
-        api_call "$uri" "$http_header" "$lyrics_file"
+        api_call "$uri" "$lyrics_file"
         echo "$REPLY" > "$api_pid_file"
-    fi
-
-    # @TODO: No idea how to get rid of the redundant-ish, first check.
-    #        Needs refactor.
-    if [[ ! -f "$api_pid_file" && -f "$http_header" ]]; then
-        cleanup_http_header "$http_header"
-        local did_err=$(cat "$http_header")
-        if [[ -n "$did_err" ]]; then
-             echo -e 'No lyrics found.\n' > "$lyrics_file"
-        fi
-        rm -f "$http_header"
     fi
 }
